len code
Chỉnh sửa trùng lặp ctsp thì tăng số lượng. Thêm add nhanh thuộc tính ở bảng add sp. Thêm bảng thương hiệu

+++ src/main/java/com/example/Nordic_SD118/entity/ChiTietSanPham.java
@@ -21,7 +21,7 @@ public class ChiTietSanPham {
     private Integer idProductDetail;
 
     @Column(name = "don_gia")
-    private BigDecimal donGia;
+    private Integer donGia;
 
     @Column(name = "so_luong_ton")
     private Integer soLuong;
